+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -10,6 +10,7 @@ import Ellipsis from "~/assets/images/icons/ellipsis.svg?react";
 import Eye from "~/assets/images/icons/eye.svg?react";
 import LeftArrow from "~/assets/images/icons/left-arrow.svg?react";
 import LeftDoubleArrow from "~/assets/images/icons/left-double-arrow.svg?react";
+import Notifications from "~/assets/images/icons/notifications.svg?react";
 import Pencil from "~/assets/images/icons/pencil.svg?react";
 import Project from "~/assets/images/icons/project.svg?react";
 import RightArrow from "~/assets/images/icons/right-arrow.svg?react";
@@ -30,6 +31,7 @@ const iconNameToSvg: Record<IconName, FC<React.SVGProps<SVGSVGElement>>> = {
 	eye: Eye,
 	leftArrow: LeftArrow,
 	leftDoubleArrow: LeftDoubleArrow,
+	notifications: Notifications,
 	pencil: Pencil,
 	project: Project,
 	rightArrow: RightArrow,