+++ apps/frontend/src/modules/contributors/libs/types/types.ts
@@ -1,4 +1,6 @@
 export {
 	type ContributorGetAllItemResponseDto,
 	type ContributorGetAllResponseDto,
+	type ContributorPatchRequestDto,
+	type ContributorPatchResponseDto,
 } from "@git-fit/shared";