+++ packages/shared/src/index.ts
@@ -5,6 +5,7 @@ export {
 	ContentType,
 	ExceptionMessage,
 	PermissionKey,
+	ProjectPermissionKey,
 	ServerErrorType,
 	SortType,
 } from "./libs/enums/enums.js";
@@ -18,8 +19,10 @@ export {
 	checkHasPermission,
 	configureString,
 	formatDate,
+	formatRelativeTime,
 	getDifferenceInDays,
 	getRelativeDate,
+	getStartOfDay,
 	initDebounce,
 	subtractDays,
 } from "./libs/helpers/helpers.js";
@@ -59,6 +62,9 @@ export {
 	ContributorError,
 	type ContributorGetAllItemResponseDto,
 	type ContributorGetAllResponseDto,
+	type ContributorPatchRequestDto,
+	type ContributorPatchResponseDto,
+	ContributorPatchValidationSchema,
 	ContributorsApiPath,
 } from "./modules/contributors/contributors.js";
 export {
@@ -78,6 +84,14 @@ export {
 	type GroupUpdateResponseDto,
 	groupUpdateValidationSchema,
 } from "./modules/groups/groups.js";
+export {
+	type NotificationBulkCreateRequestDto,
+	type NotificationBulkCreateResponseDto,
+	type NotificationCreateRequestDto,
+	type NotificationGetAllItemResponseDto,
+	type NotificationGetAllResponseDto,
+	NotificationsApiPath,
+} from "./modules/notifications/notifications.js";
 export {
 	type PermissionGetAllItemResponseDto,
 	type PermissionGetAllResponseDto,