+++ packages/shared/src/modules/notifications/libs/types/notification-create-request-dto.type.ts
@@ -1,0 +1,6 @@
+type NotificationCreateRequestDto = {
+	payload: string;
+	receiverUserId: number;
+};
+
+export { type NotificationCreateRequestDto };